Separated writer_base into the writers module again
